rename packages to reflect project name
