+++ src/main/java/sysc4806/graduateAdmissions/model/AddressBook.java
@@ -1,4 +1,4 @@
-package com.myApp.model;
+package sysc4806.graduateAdmissions.model;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Entity;
