+++ src/main/java/sysc4806/graduateAdmissions/model/BuddyInfoRepository.java
@@ -1,4 +1,4 @@
-package com.myApp.model;
+package sysc4806.graduateAdmissions.model;
 
 import java.util.List;
