+++ src/test/java/BuddyInfoTest.java
@@ -1,11 +1,7 @@
 
-
-import com.myApp.model.BuddyInfo;
-
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.test.context.event.annotation.BeforeTestClass;
+import sysc4806.graduateAdmissions.model.BuddyInfo;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.springframework.test.util.AssertionErrors.assertTrue;
