+++ src/test/java/sysc4806/graduateAdmissions/SpringMVCTest.java
@@ -1,15 +1,18 @@
-package com.myApp;
+package sysc4806.graduateAdmissions;
 
-import com.myApp.controller.BuddyController;
-import com.myApp.model.BuddyInfo;
-import com.myApp.model.BuddyInfoRepository;
-
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectWriter;
+import com.fasterxml.jackson.databind.SerializationFeature;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+import sysc4806.graduateAdmissions.model.BuddyInfo;
+
+import java.nio.charset.StandardCharsets;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
@@ -33,10 +36,27 @@ public class SpringMVCTest {
         this.mockMvc.perform(get("/buddies")).andDo(print()).andExpect(status().isOk());
     }
 
-    // How do I add a payload?
     @Test
     public void testCreatePostEndpoint() throws Exception {
-        this.mockMvc.perform(post("/create")).andDo(print()).andExpect(status().isOk());
+        MediaType APPLICATION_JSON_UTF8 =
+                new MediaType(MediaType.APPLICATION_JSON.getType(),
+                        MediaType.APPLICATION_JSON.getSubtype(),
+                        StandardCharsets.UTF_8);
+
+        BuddyInfo buddyInfo = new BuddyInfo();
+        buddyInfo.setName("bob");
+        buddyInfo.setAddress("address");
+        buddyInfo.setPhoneNumber("number");
+        buddyInfo.setAge(21);
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.configure(SerializationFeature.WRAP_ROOT_VALUE, false);
+        ObjectWriter ow = mapper.writer().withDefaultPrettyPrinter();
+        String requestJson=ow.writeValueAsString(buddyInfo);
+
+        mockMvc.perform(post("/create").contentType(APPLICATION_JSON_UTF8)
+                .content(requestJson))
+                .andDo(print())
+                .andExpect(status().isOk());
     }
 
     @Test
